app/requirements.txt
Flask==1.1.4
psycopg2-binary==2.9.1
boto3==1.18.65
botocore==1.21.65
Werkzeug==1.0.1
Jinja2==2.11.3
MarkupSafe==2.0.1